wiretap key fix and 2025build

+++ contrib/db-sync.sh
@@ -11,6 +11,7 @@ do_down() {
 	while [[ $i -gt 0 ]]; do
 	    ((i--))
 	    h="${HOSTS[$i]}"
+    	    [[ -n "$1" ]] && [[ "$1" != "$h" ]] && continue
 	    echo "#${i} Syncing ${h} DOWN"
  	   rsync -ral "${h}":/sf/config/db/banned .
 	# "${h}":/sf/config/db/token "${h}":/sf/config/db/limits .
@@ -22,6 +23,8 @@ source .env_hosts || exit
 IS_DOWN=1
 IS_UP=1
 [[ "${1,,}" == "up" ]] && unset IS_DOWN
+[[ "${1,,}" == "down" ]] && unset IS_UP
+shift 1
 
 echo "PRIVATE, LIMITS and TOKEN are now taken from SYSCOPS workstation. ADM is no longer the master"
 # SYNC private/ is DANGEROUS because those files are SOUCED by
@@ -33,13 +36,14 @@ echo "PRIVATE, LIMITS and TOKEN are now taken from SYSCOPS workstation. ADM is n
 # Reverse order so that first in HOSTS has master priority
 #rm -rf banner
 [[ -n "$IS_DOWN" ]] && {
-	do_down
+	do_down "$@"
 	echo "==[DOWN done. Press Enter to start UP]=================================================="
 	read 
 }
 
 i=0
 for h in "${HOSTS[@]}"; do
+    [[ -n "$1" ]] && [[ "$1" != "$h" ]] && continue
     echo "#$i Syncing ${h} UP"
     rsync -ral --delete  banned private token limits "${h}":'/sf/config/db'
     ((i++))
